requirements.txt
# ===============================
# متطلبات ApexFlow - PDF Processing Tool
# ApexFlow Requirements - PDF Processing Tool
# ===============================

# واجهة المستخدم الرسومية - GUI Framework
# PySide6 for modern Qt-based GUI
PySide6>=6.6.0

# معالجة ملفات PDF - PDF Processing Libraries
# PyPDF2 for basic PDF operations
PyPDF2>=3.0.1

# PyMuPDF (fitz) for advanced PDF processing and rendering
PyMuPDF>=1.23.0

# PyCryptodome for encryption and decryption
pycryptodome>=3.18.0

# معالجة الصور - Image Processing
# Pillow for image manipulation and conversion
Pillow>=10.0.1

# مراقبة النظام - System Monitoring
# psutil for system resource monitoring
psutil>=5.9.0

# pywin32 for Windows-specific functions (e.g., printing)
pywin32>=306

# ===============================
# مكتبات إضافية للتطوير والبناء
# Additional Development and Build Libraries
# ===============================

# PyInstaller for creating executable files (optional for development)
# PyInstaller>=5.13.0

# ===============================
# ملاحظات التثبيت - Installation Notes
# ===============================

# لتثبيت جميع المتطلبات:
# To install all requirements:
# pip install -r requirements.txt

# لتثبيت مع PyInstaller للبناء:
# To install with PyInstaller for building:
# pip install -r requirements.txt PyInstaller>=5.13.0

# للتحقق من الإصدارات المثبتة:
# To check installed versions:
# python check_versions.py
